Use cairosvg instead of rsvg-convert
Seems that rsvg-convert has some issues with drawing markers for runs on
the easiest difficulty. It has now been replaced by cairosvg. Using a
Python library is also a cleaner solution compared to the previous
subprocess call.

The downside is that cairosvg seems to be a bit slower than rsvg-convert.